+++ ewm-main-service/src/main/java/ru/practicum/compilation/api/controller/CompilationController.java
@@ -19,8 +19,8 @@ import javax.validation.Valid;
 import javax.validation.constraints.Min;
 import java.util.List;
 
-import static prototype.Constants.FROM;
-import static prototype.Constants.SIZE;
+import static ru.practicum.constants.Constants.FROM;
+import static ru.practicum.constants.Constants.SIZE;
 
 @RestController
 @RequiredArgsConstructor
