+++ ewm-main-service/src/main/java/ru/practicum/constants/Constants.java
@@ -1,4 +1,4 @@
-package prototype;
+package ru.practicum.constants;
 
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
