+++ ewm-main-service/src/main/java/ru/practicum/event/api/controller/EventAdminController.java
@@ -18,9 +18,9 @@ import javax.validation.constraints.Min;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static prototype.Constants.DATE_FORMAT;
-import static prototype.Constants.FROM;
-import static prototype.Constants.SIZE;
+import static ru.practicum.constants.Constants.DATE_FORMAT;
+import static ru.practicum.constants.Constants.FROM;
+import static ru.practicum.constants.Constants.SIZE;
 
 
 @Slf4j
