+++ ewm-main-service/src/main/java/ru/practicum/event/api/controller/EventController.java
@@ -15,8 +15,8 @@ import javax.validation.constraints.Min;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static prototype.Constants.FROM;
-import static prototype.Constants.SIZE;
+import static ru.practicum.constants.Constants.FROM;
+import static ru.practicum.constants.Constants.SIZE;
 
 @RestController
 @RequiredArgsConstructor
