+++ ewm-main-service/src/main/java/ru/practicum/event/api/dto/UpdateEventDto.java
@@ -15,7 +15,7 @@ import javax.validation.constraints.PositiveOrZero;
 import javax.validation.constraints.Size;
 import java.time.LocalDateTime;
 
-import static prototype.Constants.DATE_FORMAT;
+import static ru.practicum.constants.Constants.DATE_FORMAT;
 
 /**
  * <h2>Новое событие</h2>
