+++ ewm-main-service/src/main/java/ru/practicum/event/api/service/EventServiceImpl.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import prototype.Constants.StateAdminAction;
+import ru.practicum.constants.Constants.StateAdminAction;
 import ru.practicum.HitDto;
 import ru.practicum.ViewStatsDto;
 import ru.practicum.category.api.repository.CategoryRepository;
@@ -50,10 +50,10 @@ import java.util.Optional;
 import static java.lang.String.format;
 import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.toList;
-import static prototype.Constants.CATEGORY_NOT_EXISTS;
-import static prototype.Constants.EVENT_NOT_EXISTS;
-import static prototype.Constants.USER_NOT_EXISTS;
-import static prototype.Constants.checkPageable;
+import static ru.practicum.constants.Constants.CATEGORY_NOT_EXISTS;
+import static ru.practicum.constants.Constants.EVENT_NOT_EXISTS;
+import static ru.practicum.constants.Constants.USER_NOT_EXISTS;
+import static ru.practicum.constants.Constants.checkPageable;
 import static ru.practicum.event.request.api.service.EventRequestService.RequestState.CONFIRMED;
 import static ru.practicum.event.request.api.service.EventRequestService.RequestState.PENDING;
 import static ru.practicum.event.request.api.service.EventRequestService.RequestState.REJECTED;
